+++ macros/feats/maneuversRally.js
@@ -8,8 +8,9 @@ if (args[0] === "on") {
   const chaMod = args[2];
   const curTemp = target.system.attributes.hp.temp;
   const damageRoll = await new CONFIG.Dice.DamageRoll(`${dice}[temphp] + ${chaMod}`).evaluate({ async: true });
-  if (game.dice3d) game.dice3d.showForRoll(damageRoll, game.users.get(options.userId));
+  await MidiQOL.displayDSNForRoll(damageRoll, "damageRoll");
   const buf = damageRoll.total;
+  ChatMessage.create({ content: `Rally granting ${buf} temp hp to ${target.name}.` });
   if (buf > curTemp) target.update({ "system.attributes.hp.temp": buf });
 } else {
   target.update({ "system.attributes.hp.temp": 0 });