+++ macros/spells/boomingBlade.js
@@ -99,7 +99,7 @@ if (args[0].tag === "OnUse") {
     // const caster = sourceItem.parent;
     const casterToken = canvas.tokens.placeables.find((t) => t.actor?.uuid === caster.uuid);
     const damageRoll = await new CONFIG.Dice.DamageRoll(`${lastArg.efData.flags.cantripDice}d8[${damageType}]`).evaluate({ async: true });
-    if (game.dice3d) game.dice3d.showForRoll(damageRoll);
+    await MidiQOL.displayDSNForRoll(damageRoll, "damageRoll");
     const workflowItemData = duplicate(sourceItem);
     workflowItemData.system.target = { value: 1, units: "", type: "creature" };
     workflowItemData.name = "Booming Blade: Movement Damage";